src/sis/studentinfo/CourseSessionTest.java
package sis.studentinfo;

import java.util.*;
import junit.framework.TestCase;

public class CourseSessionTest extends SessionTest {
    
    public void testEndDate() {
        Date startDate = DateUtil.createDate(2003, 1, 6);
        Session session = createSession(createCourse(), startDate);
        Date sixteenWeeksOut = DateUtil.createDate(2003, 4, 25);
        assertEquals(sixteenWeeksOut, session.getEndDate());
    }
    
    public void testCount() {
        CourseSession.resetCount();
        createSession(createCourse(), new Date());
        assertEquals(1, CourseSession.getCount());
        createSession(createCourse(), new Date());
        assertEquals(2, CourseSession.getCount());
    }
    
    private Course createCourse() {
        return new Course("ENGL", "101");
    }
    
    @Override
    protected Session createSession(Course course, Date startDate) {
        return CourseSession.create(course, startDate);
    }
}
